+++ examples/FrequencyModulation/FrequencyModulation.ino
@@ -7,14 +7,14 @@ Osc aOsc1, aOsc2;
 float modIndex = 0.2;
 unsigned long msNow = millis();
 unsigned long pitchTime = msNow;
-int pitchDelta = 2000;
+unsigned long pitchDelta = 2000;
 
 void setup() {
   Serial.begin(115200);
   sineTable.sinGen(); // allocate and fill wavetable
   aOsc1.setTable(sineTable); // assign wave to osc1
   aOsc2.setTable(sineTable); // assign wave to osc2
-  // seti2sPins(38, 39, 40, 41); // BCK, WS, DOUT, DIN
+  // seti2sPins(16, 17, 18, 21); // BCK, WS, DOUT, DIN
   // useInternalDAC();
   audioStart();
 }
@@ -22,7 +22,7 @@ void setup() {
 void loop() {
   msNow = millis();
 
-  if ((unsigned long)(msNow - pitchTime) >= pitchDelta) {
+  if (msNow - pitchTime >= pitchDelta) {
     pitchTime += pitchDelta;
     float pitch = rand(36) + 48;
     aOsc1.setPitch(pitch);
