+++ examples/Reverb/Reverb.ino
@@ -14,7 +14,7 @@ unsigned long msNow = millis();
 unsigned long noteTime = msNow;
 unsigned long envTime = msNow;
 unsigned long delTime = msNow;
-int noteDelta = 1000;
+unsigned long noteDelta = 1000;
 int scale [] = {0, 2, 4, 5, 7, 9, 0, 0, 0, 0, 0};
 int reverbLength = 900; // 0 - 1024
 int reverbMix = 400; // 0 - 1024
@@ -48,7 +48,7 @@ void loop() {
   // audioLoop(); // required for internal DAC mode
   msNow = millis();
 
-  if ((unsigned long)(msNow - noteTime) >= noteDelta) {
+  if (msNow - noteTime >= noteDelta) {
       noteTime += noteDelta;
     int p = pitchQuantize(random(25) + 48, scale, 0);
     osc1.setPitch(p);
